pyproject.toml
[project]
name = "mic2e-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chat2edit>=3.0.17",
    "fastapi>=0.124.0",
    "httpx>=0.27.0",
    "numpy>=2.2.6",
    "pillow>=10.0.0",
    "python-dotenv>=1.2.1",
    "scipy>=1.15.3",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "ipykernel>=7.1.0",
    "isort>=7.0.0",
]
